backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
supabase==2.0.4
python-dotenv==1.0.0
openai==1.3.7
chromadb==0.4.18
PyMuPDF==1.23.8
google-generativeai==0.3.2
requests==2.31.0
httpx==0.25.2
pydantic==2.5.0
email-validator==2.1.0
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.1.2
aiofiles==23.2.1
pillow==10.1.0
pypdf==3.17.1
langchain==0.0.350
langchain-community==0.0.4
sentence-transformers==2.2.2
numpy==1.25.2